* CRI-pips: (22 commits)
  Removed another remaining spurious debug printing
  Remove forgotten temporarly debug statement
  Forgot to compile before previous commit :o)  Just a typo
  Fix some syntax tag that was unhanlded and lead to a pips_user_error. Thanks Béatrice for the report !
  Fix handling of perfect loop nest with enclosing block/sequence
  Yes, sometimes in PIPS you can have block statement with "NULL" as sequence...
  Allow detection of perfect loop nest even when the body of the outer loop is a sequence and the inner loop is thus alone in a sequence.
  Get rid of enum dma_dimension and switch to integer. Why do we want to limit the number of dimensions WRT to declared property ?
  Add possibility to remove an entity from the declarations in a compilation unit. This will both remove the entity from the list of entities in "code_declarations", and from statements that would hold it. This may lead to empty continue statement.
  add --tail option to adjust length. default set depending on number of processors.
  Add a new pass : set_return_type_as_typedef
  Minor fix in html prettyprinter aka IR navigator !
  svnadd of file return.c had been forgotten:-(
  handling of multiple returns within one function according to property C_PARSER_RETURN_SUBSTITUTION
  new supporting functions added to support the handling of multiple return statements in the C parser
  new property C_PARSER_RETURN_SUBSTITUTION added to cope properly with multiple goto statements in one function
  space cleaning
  Fix little mistake in IR nav !
  Add pointer basic type to html_prettyprint aka IR navigator
  Improve array_to_pointer readability by renaming previously obfuscated variable name, and fix a bug related to cast at call site.
  ...

+++ packages/PIPS/pips/src/Passes/pyps/_pyps.py
@@ -190,7 +190,7 @@ class workspace(object):
 		parents        = kwargs.setdefault("parents",        [])
 		cpypips	       = kwargs.setdefault("cpypips",        pypips)
 		recoverInclude = kwargs.setdefault("recoverInclude", True)
-		deleteOnClose  = kwargs.setdefault("deleteOnClose",  True)
+		deleteOnClose  = kwargs.setdefault("deleteOnClose",  False)
 
 		if not name :
 			name=os.path.basename(tempfile.mkdtemp("","PYPS"))
@@ -260,7 +260,8 @@ class workspace(object):
 		except RuntimeError:
 			try: cpypips.close_workspace(0)
 			except RuntimeError: pass
-			cpypips.delete_workspace(name)
+			if self.deleteOnClose:
+					cpypips.delete_workspace(name)
 			raise
 
 		if not verbose:
